+++ src/scripts/create_multi_wiki_qa.py
@@ -46,6 +46,7 @@ LANGUAGES = [
     "sk",
     "uk",
     "el",
+    "bg",
 ]